[util] Validate ID's object type.
Don't allow get, update and delete unless the item type matches
expectations.

This prevents a user with access to one object type from reading or
affecting other object types it may not have access to (for ex. by
sending a request to delete an ORDER, which it has permission to do, but
with the ID of, say, a USER, which it does not have permission to
delete).