Major revision to DynamoObject structure.
Separate database components (PkSk, AutoFields) from data, by having a
separate DynamoObjectData type, which is included as 'DynamoObject.data'
as a serde flattened field, such that the end-user doesn't see the
abstraction.

This offers a major benefit of separating the idea of "written" and
"unwritten" objects, where written objects are guaranteed to have IDs,
and unwritten objects are guaranteed to not have IDs. If a request takes
a DynamoObject::Data object, we're guaranteed to ignore any ID values, and
if a request takes a DynamoObject, we're guaranteed to have ID values by
Serde parsing. As such, no more need for id_or_critical()?, since code
can choose to either deal with DynamoObject or DynamoObject::Data. When
doing util.create(...), it now accepts DynamoObject::Data and returns a
DynamoObject.

Since we use serde flattening, this change is backwards compatible for
existing data in the database, and downstream client logic, since the
resulting JSON generated from the objects hasn't changed. Rust logic,
however, needs to be migrated.