[schema/id_calculations] For IdLogic::Timestamp, use milliseconds.
To keep consistent with UUIDs, the Timestamp should be 16 digits. Since
this easily fits milliseconds, it makes sense to be more precise.

16 digits could fit microseconds, but then it would only last until
~2300. If this is needed, a separate IdLogic can be made in the future.

Given the extra space left in 16 digits for milliseconds, the first 2
digits are still guaranteed to be 00, which could actually be a hidden
benefit for ordering when IdLogic::Timestamp is mixed with
IdLogic::Uuid, since timestamp IDs would always come first.