[util] Add support for 'creation tokens'.
Which can be generated synchronously, allowing parallel calls to
`create_item` even if one item is dependant on the others' ID.

For example, when adding a parent and a child, the child cannot be added
until the parent's ID is known. Now, a token can be generated first,
which allows both parent and child to be written simultatiously.